🎈 perf(cli): cli groups
let bilibili/xhs be cli groups
