+++ spiders_for_all/database/__main__.py
@@ -1,0 +1,4 @@
+from spiders_for_all.database.cli import cli
+
+if __name__ == "__main__":
+    cli()
