+++ spiders_for_all/spiders/xhs/__main__.py
@@ -1,0 +1,4 @@
+from spiders_for_all.spiders.xhs.cli import cli
+
+if __name__ == "__main__":
+    cli()
