+++ spiders_for_all/spiders/xhs/cli.py
@@ -1,0 +1,11 @@
+import click
+
+
+@click.group("xhs")
+def cli():
+    pass
+
+
+@cli.command()
+def test_xhs():
+    pass
